fix: google.protobuf.UInt64Value does not marshal correctly in CRDs (#8570)
* fix: google.protobuf.UInt64Value does not marshal correctly in kubernetes CRDs

* Add validation tests in helm upgrade

* More descriptive changelog

* update test to use a field which can be backported

* Fix double colon

* revert test changes

* bump to fixed version

* Test reduced max

* Adding changelog file to new location

* Deleting changelog file from old location

* Change type to Uint32

* Adding changelog file to new location

* Deleting changelog file from old location

* Fix tests

* Update test comment


* Update changelog

* Fix changelog

---------

+++ projects/gloo/pkg/api/v1/options/connection_limit/connection_limit.pb.clone.go
@@ -38,9 +38,9 @@ func (m *ConnectionLimit) Clone() proto.Message {
 	target = &ConnectionLimit{}
 
 	if h, ok := interface{}(m.GetMaxActiveConnections()).(clone.Cloner); ok {
-		target.MaxActiveConnections = h.Clone().(*github_com_golang_protobuf_ptypes_wrappers.UInt64Value)
+		target.MaxActiveConnections = h.Clone().(*github_com_golang_protobuf_ptypes_wrappers.UInt32Value)
 	} else {
-		target.MaxActiveConnections = proto.Clone(m.GetMaxActiveConnections()).(*github_com_golang_protobuf_ptypes_wrappers.UInt64Value)
+		target.MaxActiveConnections = proto.Clone(m.GetMaxActiveConnections()).(*github_com_golang_protobuf_ptypes_wrappers.UInt32Value)
 	}
 
 	if h, ok := interface{}(m.GetDelayBeforeClose()).(clone.Cloner); ok {
